Fix electrum tests: read full responses (#560)
We have been getting the `"EOF while parsing a string"` error occasionally in our electrum request tests.

It seems that sometimes we don't read the full response from the TcpStream, likely because we read anything that is already on the stream without knowing if all the data is already there. So I am **trying** to fix it by reading a full line, until we reach a `\n`, which is the terminator used by electrum responses.