Added enum to better express find-who-is-lying
Before this commit an Option was being used to identify whether one of
the peers were lying or not. This made difficult to handle cases of
unresponsiveness or both peers lying. This commit adds an enum to
enhance expressability.

It also:

Updated docstring

Removed None field of enum PeerCheck

Renamed block to block_hash

Modified PeerCheck

Added suggestions from @JoseSK999