chore(wire): Add enum to better express find-who-is-lying
Before this commit an Option was being used to identify whether one of
the peers were lying or not. This made difficult to handle cases of
unresponsiveness or both peers lying. This commit adds an enum to
enhance expressability.

The peer can be either reponsive and send the correct accumulator,
`unresponsive` or `lying`. If the peer is lying, the accumulator it sent
is marked as invalid and a shutdown message is sent to disconnect.