chore: remove the `futures` crate (#564)
When I started this project, we used async-std as async runtime. async-std did lack some features, like oneshot channes and some macros like select!, for that reason, I've used the futures crate to get some of those functionalities. However, since #172, we use tokio as async runtime. And tokio does have those features. So I'm removing futures as dependencies.

tokio also has out-of-the-box support for ctr-c handling, so I'm using it instead of the crate.

I've also made the public methods in florestad async. This simplifies some stuff, and shouldn't be a problem, since these methods should only be called with a running runtime. If you need to use them in a sync environment, use a thin wrapper that owns florestad and the runtime. Then block_on every call.

* remove futures from floresta-electrum

* remove futures and ctrc from florestad and wire

* remove futures from floresta