fix: re-index didn't apply changes to chainstate (#562)
Before this commit, the re-index code would find the right place
where we would work from, but not apply it to our chain. This could
create spurious errors like #551.

This commit fixes that, and also improves how we handle accumulators
on re-indexes. We now try to find our latest accumulator, move
validation index there, nad mark future blocks as `HeadersOnly` to
not have problems with `BlockDoesntExtendTip`.