chore: remove `expect`, add error handling and lints for `florestad/src`.
Replaced `expect` methods in favor of more properly way to handle errors
like `map_err` `map_or_else` and `unwrap_or_else`. Also, remove calls of
`panic!` and `std::proccess::exit` because we can't just kill the
process every time we have a problem and this approach could redirect
the error to the caller to handle it.

follow-up #463.