Don't always rebuild core and utreexod (#580)
* prepare.sh: Always build florestad

Since floresta is what we are testing, it makes sense to always build it.
This also avoids the need to pass `--build` every time you change floresta,
which would re-build utreexod and bitcoind, even tough they didn't change.

* justfile: don't use --build for functional test

This will cause the receipt to only build core and utreexod if we
didn't already.