Remove `assume_valid_index` and bump store version
If I remember correctly this field was used to know whether to validate scripts for a given block. But now `verify_script` searches the assume-valid header and uses the stored height instead, only if it's found to be in the best chain (which is the correct approach).

So we are not populating this field nor reading its value anymore. I think we can remove it entirely as it serves no purpose.

This change is not backwards compatible for our `FlatChainStore`: we can't safely load pre-existing flat store databases because the expected `metadata.bin` length and layout is different.