fix(test_framework): kill spawned processes if tests fail due timeout error (#592)
Previously, the test_framework will not kill the spawned nodes if
we receive a timeout error during the procedure of detect random ports.
This commit fix this wrapping the `FlorestaTestFramework::run_node`
method with a `try/except` clause and, in case of error, try to kill the
node gracefully or with SIGTERM/SIGKILL if needed.

Fix #589