Remove `assume_valid_index` and bump store version (#593)
If I remember correctly this field was used to know whether to validate scripts for a given block. But now `verify_script` searches the assume-valid header and uses the stored height instead, only if it's found to be in the best chain (which is the correct approach).

So we are not populating this field nor reading its value anymore. I think we can remove it entirely as it serves no purpose.

Implemented a migration by checking if the `metadata.bin` length is the old length. Then we rename the old file to `metadata.bin.old`. For testing this we use an old metadata file.