Fix reindex during chain selection (#599)
In `reindex_chain` we don't update the best block field if there's no accumulator associated, but this must be allowed in the chain selection phase.

Another problem is that when we try to shutdown the node during chain selection we don't flush the chain, so we don't save the `BestChain`, which means the next re-start will not identify the chain as initialized and will reindex. But this reindex loops forever because we never update the `BestChain` as explained above. I fix this by calling `shutdown`.

Also added warns before each `reindex_chain` such that we know the reason in each case.