feat(wire): implement a `send_to_fastest_peer` method
Some requests may take a long time for slower peer, and don't really
require a diverse set of peers. The main expample for this is when
we download blocks: we already know what the best chain is, so
there's no problem in requiring this to the peers that can serve
it in the least amount of time.

This commit implements a function that looks at the historical delay
of our peers in replying to our messages, and take the smallest one
to send a request.