wire: use `send_to_fastest_peer` in `chain_selector.rs`
This commit replaces `send_to_random_peer` with `send_to_fastest_peer`
where appropriate in `chain_selector.rs`.

Note that it is ok to use this during the headers sync, because at the
end, we will ask our peers for their headers, which should show us
if the fastest peer is trying to do something malicious.