wire: allow sync node to make more block requests
Since now we are not asking for blocks from the same peers, we can
increase the amount of inflight requests we can sustain. I'm increasing
the number of `getdata` requests to 10, and each one can have at most
5 blocks, for a total of 50 blocks. At a 4MiB maximum, that's a 200MiB
limit of data (actually a little more due to them being parsed, but still
manageable).