feat(floresta-chain): add option to subscribe to blocks with spent UTXOs (#596)
Introduces a new subscription mechanism that provides both block data and input UTXO data to subscribers. This enables applications that require knowledge of the spent UTXOs.

Changes:
  - New trait: BlockWithInputsConsumer for subscribers that need input data alongside blocks
  - New method: Use subscribe_with_inputs() to subscribe to block data with input UTXOs
  - Backward compatible: Existing BlockConsumer subscribers continue to work unchanged

Use Cases:

The primary motivation for this PR is to support metaprotocol indexers, which require access to the spent UTXOs of each transaction, but not the full UTXO set. Developers can then integrate Floresta directly into their indexer, so that users do not need to run a separate full node in order to index the chain.

In addition, this PR enables basic blockchain analytics on light clients. Light clients can track recent fee rates and other transaction patterns that may be useful to the user.

Finally, this PR enables light clients that track spends from addresses with unknown balances, without needing to download filters or reindex the entire chain. For example, as an intrusion detection mechanism, a user could setup a watchtower that notifies them of spends from their wallet, providing only their wallet's descriptor.