refactor(integration): use random port instead of parsing it from logs
Currently in our integration tests, in order to have an unique port on
each daemon we run, we pass `0` as port, which by convention
means "pick any available port". However, our log file parsing is
a bit brittle, and sometimes it misses the port log. This causes
our tests to think our daemon didn't start, where it actually did.

This commit changes this behavior by picking a random port in
the upper-range, and passes it to the starting daemon. Since
we pick a random port out of thousands (usually assigned to
outgoing sockets), the likelihood of collision is insignificant