[Feat] Removing instalation from default pipeline.
The spell-check recipe was installing the typos binary by default. This
commit changes it and delegate to the user to provide typos, the recipe
now relies on the check-command recipe to fail wheter a certain binary
isnt present offering a link that should tell how to get it when needed.