RPC: don't use the [] syntax, as it will panic
Using brackets for getting an item from an array is a bad practice,
because if the element isn't there, we will panic. This commit
introduces a new way to handle parameters, with actual error
handling for missing parameters.

I've also introduced a few helper functions to avoid code duplication.

Finally, I've added a RpcRequest struct to avoid using the black box
"Value" type around, which require error handling all the time we
access it.