Apply suggestions from code review fixing unreachable code
Co-authored-by: Davidson Souza <me@dlsouza.lol>

Improved syntax for checking if the received block had not been
modified.

Fixed the logic checking of adding a peer acc to the list of invalid
accumulators by continuing the loop after it instead of fininshing it.

Added suggestions from code review

Removed unnused variable _acc2

Renamed _filters to filters and _headers to headers

Modified logic to find the headers corresponding to the given hashes.
Instead of using a position variable, now it uses the method .find().