Add check on whole acc
Previous this commit, the accumulator was being partially checked.
It had a check on only one of the leaves. Now, using acc_hashes! macro
it builds a whole acc from the given strings, thus enabling correctly
checking the accumulator.

Also removed unused code, moved imports to module level and renamed
headers_iter to invalid_filters_iter.

Replaced value of number of blocks for cnst. Previously it used a
hardcoded value of 119, which was different then the available 120
blocks.

Replaced assert to consider all the blocks available. Previously it was
cutting short of one block (119 instead of 120).

Removed outdated comment.