wire: Use hardcoded peers if we can't find addresses
If we don't have an address with the desired services, we may just
return a "NoAddressesAvailable" error, but not do anything about it.
We currently only create new connections if we pass more than one minute
without any peer. However, sometimes we can create connections with peers
given by dns seed, but not utreexo ones. Thus causing the node to stay stuck
during IBD.

This commit adds the hardcoded peers to our address manager if we get a
"NoAddressAvailable" error as a fallback mechanism.