florestad: fix the version string for consistency (#640)
Before this commit, if we build from a git repository, the user string would
be something like "v0.8.0". However, if you build from an archive, it would be
"0.8.0".

The user agent is also different: "/Floresta/v0.8.0/" and "/Floresta/0.8.0/"
respectively. This commit makes sure the version always have the 'v' prefix, but
user agents don't, as user agent's versions are usually numerical.