feat(node): subscribe to logging event with `tracing-subscriber`
Previously, logging was done with the `log` and `env_logger` crates.
For better DX on downstream projects, we use the `tracing` crate
to emit logging events on all of Floresta's crates, and implement
a subscriber for these events on `crates/floresta-node/src/florestad.rs`.
Downstream projects should also implement a subscriber to consume these
logging events.