feat: move logging to use `tracing` and `tracing-subscriber` (#643)
* feat(florestad): swap `log` for `tracing`

* feat(chain): swap `log` for `tracing`

* feat(electrum): swap `log` for `tracing`

* feat(watch-only): swap `log` for `tracing`

* feat(wire): swap `log` for `tracing`

* feat(node): subscribe to logging event with `tracing-subscriber`

Previously, logging was done with the `log` and `env_logger` crates.
For better DX on downstream projects, we use the `tracing` crate
to emit logging events on all of Floresta's crates, and implement
a subscriber for these events on `crates/floresta-node/src/florestad.rs`.
Downstream projects should also implement a subscriber to consume these
logging events.

* chore: change log file name from `output.log` to `debug.log`

* chore(florestad): change `stopping florestad...` log message to `stopping Floresta`