Move logging subscriber to `bin/florestad`
I have moved the `data_dir_path` function to `main.rs`, as we need it to set up the logger (file log), and its tests.

I am also creating the directory in the binary, and the library code only receives the path and validates the directory.

So `Florestad::Config` now requires a value for the `data_dir`. `Florestad::new` now requires at least a `data_dir` and `Network`, and I removed `Florestad::default()`.