doc: Update PR template to remove unneeded fields
Our PR template is too verbose and has useless fields. Once we've
started using the merge-script, our merge messages are
unnecessarilly big. This commit removes the intention and crates
changed fields, as those should be in the PR tile.

It also makes it required to remove the checklist when creating the
PR