chore(nix): remove nix packaging and simplify nix code.
We decided that packaging should be done outside the main repo, this commit
deletes the nix packaging.

Another change being introduced is the simplification of the nix code, it was
a choke point of the project and sometimes it was bringing problems to some
PRs such as the one that introduced bitcoinkernel.
That said, everything nix-related that may block merging prs because of changes needed
on the nix side was removed, including the checks, theyre now on the devshell.

The devshell hooks now are responsible to apply the previously removed
checks that was already adopted in our CI and CONTRIBUTING guide.

This commit present changes to make everything compliant with all these checks.

This commit broke how the latter devshell
for running the python tests worked. This commit
fix it using the same build derivation as before
but from the externally managed floresta-flake.

The python-deps devenv ends up being more concise,
integrating only things related to this finality,
one that wish to have the all the checks and should
other related project dependencies should enter the
default devshell too.