chore(nix): remove nix packaging and simplify nix code.
We decided that packaging should be done outside the main repo, this commit
deletes the nix packaging.

Another change being introduced is the simplification of the nix code, it was
a choke point of the project and sometimes it was bringing problems to some
PRs such as the one that introduced bitcoinkernel.
That said, everything nix-related that may block merging prs because of changes needed
on the nix side was removed, including the checks, theyre now on the devshell.

The devshell hooks now introduces the previously removed checks that was already
adopted in our CI and CONTRIBUTING guide.

This commit is compliant with all these added checks.