test(wire): Add tests to catch lying peers during PoW fraud proofs
This commit adds tests to check if the node can get to the correct
accumulator even if only one peer is being honest.

The test data was taken from a utreexod node running on signet.

The test is executed by mocking peers with fake accumulators at
different block heights. During sync, the Floresta Node is connected to
only one honest peer and checks each accumulator received until it has
only one accumulator and a true accumulator.

Removed hashes.json as it is not needed anymore. It will now use the
block and calculate the hash.