feat(florestad): enable `assume-valid=0`
Previously our `--assume-valid` flag only took an option, and `None` was treated as the hardcoded assume-valid value.

I have instead made `florestad` parse an `AssumeValidArg` directly, so that users can specify `--assume-valid 0`.

Now you can also pass `--assume-valid hardcoded`, which is the default so not needed. You can still pass a custom hash as before, but it's validated in the command-line parser.