feat(functional): introduce pytest and convert `/example`
Introduce pytest as the new testing framework to simplify and enhance the way
tests are written and executed. Pytest facilitates the creation of reusable
functions and fixtures, making it easier to manage and instantiate nodes
dynamically. It also provides built-in assertion functions and a robust
test runner.

All tests in the `/example` directory have been migrated to use the pytest
framework.

co-authored-by: j-moreno-c-r <j-moreno-c-r@users.noreply.github.com>