fix(consensus): check for amount overflows
Our consensus checks are adding satoshi amounts without checking overflows, and so it's actually possible to create a transaction with huge output amounts that passes as valid since an overflow makes the `out_value` be small. In that case the attacker creates coins out of thin air, violating the 21M limit, and we think it's a valid transaction. I've fixed this kind of errors by calling `checked_add`.

I've also moved the validation logic that doesn't require the spent outputs to a new function called `check_transaction_context_free`. These checks are:
- At least 1 input and 1 outputs
- No null prevouts (non-coinbase)
- Valid input script sizes
- Sum of output values is < 21M

This new function is useful as it can be called for assume-valid SwiftSync (TODO) where we don't have the spent UTXOs  at validation time.